move submits one level down
